backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1

# Server
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# PDF Processing
PyPDF2==3.0.1

# AI Integration
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Vector DB
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
Pillow==10.2.0
